I looked into eMMC behavior in depth today because I had an external card that was exhibiting gruesome errors. After a lot of mucking around I know a lot more about the linux mmc system. However nothing I changed did got rid of the errors. My conclusion is that our settings were pretty much ok and the external card is just damaged.
In reading the driver documentation, the omap-hsmmc driver in linux defaults to a max speed of 52MHz. The am335x MMC controller operates at “up to” 48 MHz and "1, 4 or 8 bits”.  48 MHz is specified as the max frequency in the datasheet. This is 2 * 24 MHz MCLK, reasonable. Since the kernel is being told max_freq is 52 Mhz this implies it is being overclocked. I though I found something important here. However, after changing the device tree and seeing no change in write performance I read the TRM in more detail and now I do not think it is possible to adjust the clock speed to anything other than 48 MHz. (That is an oversimplification: it starts up very slow and in 1-bit mode and negotiates up to “high speed”, and “high speed” will be 24 MHz instead of 48 MHz in the low-power mode opp50, but you cannot micro-adjust the frequency in software like you can with the DRAM subsystem.)

So the standard am335x device tree (that we use along with TI am335-evm and beaglebone etc) is telling the kernel to do something stupid, namely set the clock faster than it can really go, but it does not have effect on the actual clock speed. Nevertheless,  I am changing the device tree max_freq to 48 MHz so that the kernel is being told the truth about clock speed.

* I am also telling the kernel to disable the write-protect feature (we already disabled the card detect feature)
* I determined we should use “ti,am33-hsmmc” instead of “ti,omap4-hsmmc”.

However the difference in the last setting involves exactly one register setting that works around a bug in the am335x when waking the device from sleep, which we never do. So it doesn’t matter either.